tor-dirserver: Implement download timeout
This commit starts work in the `mirror::operation` module by
implementing the functions for calculating the `Duration` to wait from a
given `SystemTime` until the point in time when the dirmirror should
download new documents from the authorities again.

Also, this commit introduces new error types and internal helper
functions in order to fulfill this purpose better.