tor-hsservice: Fix time_store test.
This behaviour was changed in humantime 2.3.0. Because they made this
breaking change on a minor semver bump, if we want to depend on this
behaviour, we would need to lock to a specific version. I don't think
this is critical, but I am still looking into where exactly this is
used.