tor-netdoc: Mirror netdoc structure in RouterStatus document digests
In md consensuses the referenced document digest is in the `m` field.
This seems to be to make it similar to votes.  But really the function
of the md desc hash in md rs entries is much more parallel to the
plain desc hash in plain rs entries.  The way the spec has done it
means the `r` item has a needlessly different syntax.