ci: Move maint-check-changelog into maint-checks.
I hope that this will reduce the load on our CI and speed up these
checks. This should also cause this check to run earlier in the process,
which is nice for finding issues earlier.

It's unclear to me if there was a specific reason this was a separate
job. As far as I can tell from the commit message and discussion in
!2126, there does not seem to be a specific reason.