ci: Move "check" jobs into "build" stage.
This moves everything that was in the "check" stage into the "build"
stage, so that beginning to build the artifacts isn't blocked by running
the checks. This should result in CI taking less time overall.

There is some concern that making CI more efficient by increasing
parallelism could increase load on runners and cause more problems than
it solves, so that's something we should look out for in evaluating this
change.