tor-netdoc: Explain that Ignored can't be used as an argument
It is FromStr (which is fine).  Which emant it was
ItemArgumentParseable but but it can't be Display so it can't be
NormalItemArgument.  This illuminates the fact that it shouldn't
really be ItemArgumentParseable since we're going to want all types
that are in netdoc structs to be both parseable and printable.