tor-netdoc: parse2: Work around MSRV unhelpful lifetime
1.85.1 rejects this, because it doesn't extend the lifetime of the
`format_args!`.  1.90 is fine with it.

I want to keep `format!` not `String::new()` so that it's easy to
change back.  That means a clippy allow.  Getting that to apply to the
code inside the derive is tricky, so I've enabled it for the whole
module.  (IMO this clippy lint is pretty annoying anyway.)