CI: split rust-recent script and jobs
rust-recent (and rust-latest, for runs on the main branch) is one of our
longer jobs, particularly in the 'build' CI stage, making it a
bottleneck for starting the 'test' stage.

The job does quite a bit of work, some of it with very little overlap
due to use of different cargo build profiles.

This commit splits the job roughly by build-profile, so that the work
done with each build profile can proceed in parallel.