CI rust-recent-dev-test: split out test
This job is still expensive and a bottleneck. "cargo test" is the
slowest part, and experimentally doesn't seem to generate artifacts used
in the other steps nor reuse artifacts genrated in other steps.
Splitting it out to its own job exposes more parallelism.

"cargo test" is also more prone to flakiness than the other commands,
which just build code without running it. Isolating the flakiness is
helpful when we need to retry. (Though of course ultimately we want to
eliminate such flakinesss)