tor-dirserver: Outline the dirmirror operation
This commit creates a function named `mirror::operation::serve`, that
will serve as the primary entry point for the dirmirror operation.

While the function itself already contains some logic, such as querying
the timeout to wait for the next consensus, it is largely a rustdoc
design document that adds lots of comment to the `mirror::operation`
module.