tor-netdoc: Add a lot of spec cross-references and tidy docs
Add links to torspec, and other cross-references.

Change heading fof `IgnoredPublicationTimeSp` to `publication` to
match spec and distinguish it from the `published` item in netstatus
docs.

Align several descriptions with the wording from torspec.

Remove a restatement of a fact from the spec, on the `valid_until`
field in `Lifetime`.