relay: Implement a blocking run() that starts the relay
We have bootstrap() to prepare everything we need before starting the
relay which yields a TorRelay.

This commit then adds a run() function which starts the relay by
launching background task. At the moment, we only have one task and it
is the channel expiry one.

The relay blocks on the task handle set until one task is on error or
stops which leads to shutting down the relay.

Part of #2217

Signed-off-by: David Goulet <dgoulet@torproject.org>