ci: Add needs to almost all jobs.
This also updates some needs statements that existed to order jobs that
did not actually have a dependency on each other.

The idea here is to run everything with as much parallelism as possible,
and if that parallelism causes problems, we should ideally solve it by
adding more runner capacity, or taking a closer look at what the actual
problem is.