arti: Split socks-specific parts out of handle_socks_conn
We want to abstract every part of this code that knows that it's
speaking SOCKS, so that we can in turn add a new proxy type.

Note that several methods and types here have names that are no
longer appropriate for their functionality.  I'll revise those in a
later commit.