tor-dirserver: Make database full sync
This commit makes the entire database operations synchronous, thereby
replacing deadpool with r2d2.

The full motivation is outlined in a rustdoc comment at the top of the
`database` module, but it can be summarized to the fact that SQLite is
by design inherently synchronous due to directly interfacing with the
file system.