arti(-client|-ureq)/tor-hsservice: semver files v2
Combined with the previous commit, this completes my second attempt at
correctly handling the semver files.

tor-hssservice:
I removed the extra bit at the end of the first line explaining the
semantics of the new return type, since the important part is really
just the fact that the return type changed. The semantics are documented
in the methods themselves.

arti-client:
Same as above, but I also removed the `client` module prefix since the
`client` module isn't actually public, so to external crate users the
broken type is directly in the crate root.

Removed old "used in the public API of this crate" line.

Added lines referring to the 2 breaking changes we re-export from
tor-hsservice.

arti-ureq:
Removed old "used in the public API of this crate" line.

Added lines referencing every breaking change in arti-client. This MIGHT
be wrong, but I think this is correct because we re-export arti-client
as a whole.

Signed-off-by: hashcatHitman <3924-hashcatHitman@gitlab.torproject.org>