tor-dirserver: bump MSRV from 1.83 to 1.86
I noticed that when doing `cargo generate-lockfile` in the workspace
root, it was locking dependencies to 1.83 compatible versions. It turns
out tor-dirserver started on MSRV 1.83 after we already bumped to 1.85,
and has been missed in MSRV updates ever since.

I've found no indication it is intentionally getting this special
treatment and am assuming it is a mistake.

Signed-off-by: hashcatHitman <3924-hashcatHitman@gitlab.torproject.org>