tor-dirserver: Apply database changes to http
This commit applies the recent changes to the database module into the
http and http::cache module.

Notably, this includes the use of the database::sql! macro as well as
the use of database::read_tx and database::rw_tx for creation
transaction and saving, as well as retrying, although the latter one is
subject to SQLite's own busy handler