proto: Add RelayIdentities object holding our keys
This is a intermediary object between tor-chanmgr and tor-proto that is
when building a relay channel, those keys/certs need to be set in the
ChannelBuilder so the tor-proto can use them to authenticate.

We avoid that way making tor-proto depending on tor-keymgr for the
ultimate goal to avoid tor-proto to have access to all the keys in the
KeyMgr.

Future commits will introduce a relay channel builder which will use
that object to set the keys.

Signed-off-by: David Goulet <dgoulet@torproject.org>