proto: Compute authenticate data after validating certificates
This is only done if we kept the AUTH_CHALLENGE cell and we have relay
identities. In other words, this is only when the UnverifiedChannel was
created from a RelayInitiatorHandshake.

Note: The check_internal() function is too large and should be
refactored in smaller pieces.

Note: It is also likely that we need to split UnverifiedChannel and
VerifiedChannel as it is getting client or relay members. Not great.

Signed-off-by: David Goulet <dgoulet@torproject.org>