proto: Use CircuitCellSender in the backward reactor
The backward reactor needs this for handling padding (because it may
need to e.g. put the chan sender in a blocked state).

Note that `CircuitCellSender` has an underlying
`SometimesUnboundedSink`, so from now on we must be very careful and
avoid writing to the `chan_sender` if it's not ready (because otherwise
we can end up buffering unboundedly). The next commit will add some
extra checks against reading from futures::Streams that might cause us
to write to the `chan_sender` sink (futures::Sink).