proto: Add UnverifiedRelayChannel implementation
This commits takes out the relay specific code out of UnverifiedChannel
and puts it in UnverifiedRelayChannel.

In order to pull this off, we added some fiels to the VerifiedChannel
struct so the relay code get back generated data from the
UnverifiedChannel::check() in order to do its validation work and yield
a VerifiedRelayChannel.

This also lead to a cleanup of expect() and ChannelType::is_initiator().

Signed-off-by: David Goulet <dgoulet@torproject.org>