tor-dirserver: Implement read_tx and rw_tx
This commit does a large overhaul in the overall way on how we interact
with SQLite.

Although SQLite is a fairly straightforward tool, it has one large
caveat, namely the handling of the infamous `SQLITE_BUSY` error.

In WAL mode, the journaling mode we are using, the respective error may
occur if there is an existing write transaction while another thread is
trying to either upgrade an existing read transaction or create an
exclusive transaction at the same time.  In this case, SQLite will fail
at the statement that either requires the write operation or at the
creation of the transaction respectively.

In order to tackle this problem, SQLite provides the `busy_timeout`
pragma, which allows specifying a timeout in milliseconds during which
SQLite will retry to obtain a write transaction or fail immediately.

This only works in the latter case, that is, creating a write
transaction from the beginning.

To solve this issue, we introduce two new functions:
1. `database::read_tx`
2. `database::rw_tx`

Both functions accept a database pool alongside a closure accepting a
`Transaction` and returning an arbitrary return value.

The first function creates a deferred transaction that gets rolled-back
in the end, making it suitable for read-only connections.

The latter function creates an exclusive transaction that gets committed
in the end, making it suitable for read-write connections.  This also
honors the `busy_timeout`, i.e. the function retries for up to 1s to
acquire a write transaction before failing ultimately.