circmgr: Implement expiration for long-lived tunnels.
This code uses the algorithm of prop368: When a tunnel is
sufficiently isolated, we want to expire it after it has been
disused for a certain amount of time.

The implementation is a little tricky, since we need to call an
async function on ClientTunnel to look up the disused_since() check.
We don't want to call async functions while holding locks, so
we need to drop the TunnelList lock before we do the disused_since()
calls, and the grab it again.

As part of this change, I've had made expiration functions return
the earliest time at which any tunnel might expire.  We can use
this in the future to spawn fewer expiration tasks, and scan the
list of tunnels less often.