proto: Make StreamTarget take a generic Tunnel
This is needed because we want to reuse `StreamTarget` and `DataStream`
on the relay side too, but to do that, we need to abstract away the
tunnel/circuit type (prior to this MR, `StreamTarget` was was
client-specific, as it used to wrap a client tunnel).

Note that `StreamTarget` needs a handle to the client/relay circuit
reactor because it needs to be able to shut down the circuit if a
protocol error occurs (cells carrying stream data are parsed late,
*outside* of the reactor, so if e.g. a cell fails to parse, the
`DataReaderImpl` needs to be able to shut it down), and because it needs
to be able to inform the reactor of flow control-related events (such as
drain rate update).