proto: Make cc state shared between CircHop{Inbound,Outbound}
The relay reactor needs the cc state to be shared between the inbound
and outbound components (i.e. the forward and backward reactors), so we
need to put the cc state behind a mutex.

There will never be any contention on this mutex in the client impl,
because the client reactor doesn't split the `CircHop` into
`CircHopInbound` and `CircHopOutbound`.

In the future, we should work on trying to reduce the number of locks in
the `CircHop` states.