tor-netdoc: Implement KeywordEncodable for &'static str
My planned encoding derive macro is going to provide keywords as
literal strings.

(The purpose of this trait is to allow the old parser to be generic
over its keyword constants, not to provide type-safety against passing
non-keyword strings to document encoders.  Such bugs aren't very
likely and will be especially unlikely when we're deriving encoders.)