RelayFlags: Use the singular RelayFlag for naming individual flags
We're going to separate RelayFlag from RelayFlags.

We could continue to provide a bunch of constant values for
RelayFlags, for each individual type, but that would involve an ad-hoc
derive, and would still be a bit of an unusual API.

Instead, we're going to make ``RelayFlag` a normal enum.

In this commit, we change everyone to refer to it by its singular
name.  Hopefully separating out this bulk change makes both this, and
the core commit, easier to review.