tor-netdoc: Replace RelayFlag with an enum and RelayFlags with an EnumSet
Now:

`RelayFlag` is a single flag, which is a different type to
`RelayFlags`.

`FromStr` and `Display` are now implemented in a normal way for a
single flag.

We don't use bitflags, with its funky input syntax, any more.
Instead, we have a normal enum.

No longer is it possible to have a `RelayFlags` with "unknown bits"
set (whatever that means).  (Therefore, an associated test case for
this is now gone.)

The names are principally specified just as enum variant names.  The
enum variant names can be identical to the network document keywords.

Briefly, we provide constants in capitals, for the benefit of all the
call sites.  That means that *this* commit can contain only the
primary changes.

A few macro-generated things have been adjusted in this commit, but
they are actually going to go away.

Indentation is now wrong.  We'll fix that at the end.