proto: Start handling incoming stream requests
The relay reactor is now able to handle incoming stream requests (i.e.
cells that open streams). It currently only supports DATA stream
requests (BEGIN); support for other stream types (BEGIN_DIR, RESOLVE)
will be added later.

`Reactor::new()` now returns the futures::Stream of Tor streams, alongside
the `Reactor` and `RelayCirc` handle. Whoever calls `Reactor::new()` is
responsible for passing the stream of streams over to the task that is
meant to handle it ("handle" in this case means either rejecting the
stream with a given `END` cell, or accepting it and forwarding the
connection between it and the corresponding application stream).

IMPORTANT: the above is a bit half-baked! Next on my TODO list is is to
iron out the details of how/where this will actually be handled.
I am also a bit unsure about the API here: I think it might've been
nicer to give the user the ability to obtain this `futures::Stream` from
`RelayCirc`, which is, after all, a handle to the reactor?
Also on my short-term TODO list is to figure out how conflux will affect
this API and usage.

And there is another wrinkle here: for incoming DATA stream requests,
the handler will need to produce a resulting `DataStream`, which is not
yet fully implementation-agnostic (it wraps a `ClientDataStreamCtrl`).
This too will be handled in a separate MR.