doc/hsc.md: Remove note about invalid keystores
I don't think we need to tell users a C Tor keystore cannot have more
than one client entry for the same onion service, because they already
know that: C Tor doesn't allow it, so if they are migrating an existing
C Tor configuration, it will likely be valid. And if not, the error
message will inform them what went wrong.

So I am removing this section in the interest in keeping this guide
short and sweet and straight to the point.