Take `self` rather than `&self`
This commit modifies `SaturatingTime::saturating_{add, sub}` to take
`self` rather than `&self` as this is the behavior of the Rust standard
library.

Internally, this also adds `Copy` as a requirement for the internal
trait, which is satisfied by `SystemTime` and `Instant`.

Also, the `internal::SaturatingTime::checked_{add, sub}` continue to
take `&self` as this aligns better with the behavior of the Rust
standard library.

Fixes #1