tor-dirserver: Improve database::Timestamp
This commit improves upon the work made in arti!3460, by establishing
the `database::Timestamp` type as a general purpose type for the
handling of timestamp throughout the code base.

It also replaces all previous occurrences of `SystemTime` within
`dirmirror::operation`, which leads to the reduction of code complexity
in one particular function, namely `calculate_sync_timeout`.