parse2: Move naming convention to poc
This commit moves the "Naming Convention" of the parse2::poc module into
the respective module, as it does not apply to parse2 itself, as
outlined in arti!3462.

With this commit, there remains `NdaSystemTimeDeprecatedSyntax` as the
only type outside poc following this naming schema.  However, because it
is more or less already deprecated anyways, there is no real need to
bother with it any further.