tor-netdoc: Throw rather than squirreling error from NormalItemArgument
NormalItemArgument is for types where we use the Display as the netdoc
argument formatter.  But what if gives the empty string?

Previously we would allow `add_arg` to handle the error.  That would
record it in the NetdocEncoder.  That's kind of OK, but it will
prevent the caller from aborting early (and from elaborating the error).