tor-netdoc: recast Ignored and impl new encoding traits
This is a bit subtle.  Ignored is used where we want to see and accept
a value (for example, an Item or an Object), but we don't want to emit
it.

Update the docs and implement the various traits.  We need an
uninhabited helper type which implements ItemValueEncodable and
ItemObjectEncodable but *not* ItemArgument.  That prevents the use of
Ignored for an argument, which would be wrong because what would we
emit to keep the argument parser in step?