tor-netdoc: test2: check round-trip encoding
The test case documents are often not in canonical form.  We don't
want to change them to be in canonical form because we want to test
parsing edge cases.  So, we provide a mangling scheme with in-document
notes, and annotate all the documents.

Apply suitable annotations to all the documents.

Test that the encoded forms are as expected.

This reverts commit 05e7b8b60061a7e8a8d004a822cc213d3ea22d47.