map_builder: Alternative syntax for use by derive(TorConfig)
derive_deftly wants to expand types before passing them to
macro_rules macros, which is quite reasonable.  But map_builder
wants its input collection type to be an `ident`, not a `path`.
(And macro_rules doesn't accept `path` before a `<`.)

To fix this, we're providing an alternative syntax for map_builder,
where the inputs are the map type and the builder map type.