tor-config: new derive_deftly(TorConfig) template.
This template is meant to replace most of our use of derive_builder
for configuration objects.  Where possible and reasonable, it
delegates to existing macros, and automatically infers what special
patterns we use for individual types.  In other cases, it uses
compile-time errors to inform the caller about pattern violations.