Pin our derive-deftly version to ~1.6.0 everywhere
We are supposed to pin whenever we enable the `beta` cargo feature,
see
  https://docs.rs/derive-deftly/latest/derive_deftly/doc_changelog/index.html#beta-features
Empirically, we somehow failed to do that in tor-circmgr.

In practice not pinning makes little difference since cargo wants to
pick the same version everywhere, but we should be correct.  But it is
more maintainable to pin everywhere.