check_licenses: Adjust comment on priority-queue exception
Technically `priority-queue` belongs in the `MPL_20_OK` list, but our
current checks are too rudimentary to be able to handle the `OR` (we
*do* split on `OR` for one of the checks, but we only compare the
resulting licenses against `RECOGNIZED_LICENSES`, not `MPL_20_OK`). So
for now, let's retain the exception for `priority-queue`.

(We could expand the logic in the branch handling the `OR'`d licenses to
fix this issue, but the result would be quite fragile; and besides, I
think we're going to eventually want to replace this with a proper
license parser).