proto: Client circuit reactor now handles AnyChanMsg
This commit removes the CircuitRx* based solely on the client circuit
message and moves it into the top level of the crate so all reactors can
use them.

The client reactor then upon receiving the message, it converts the
AnyChanMsg into a ClientCircChanMsg. On error, this leads to a shutdown
of the entire reactor due to a fatal error.

In order to pull this off, we added a CircuitAction::Shutdown that is
handled as a priority.

Signed-off-by: David Goulet <dgoulet@torproject.org>