shadow CI: disable stallouts for onion clients
tgen aborts a stream and starts the next one when the "stallout" timeout
is reached. In some cases the arti (or tor) client just needs a bit more
time to establish a connection to the hidden service.

A potential downside is that if a single stream "stalls" forever without
an explicit error, tgen will never give up on it, and thus never try any
more streams. We can worry about that if and when we see it though; and
it would indicate a likely bug since the known clients and servers in
this test shouldn't stall indefinitely in this way.