arti: remove 'rustls/tls12' feature
We don't rely on this within the arti crate, and tor-rtcompat already
sets this. It's confusing to have it here as well. If the arti ecosystem
requires this, it should be enabled on the lower level crates (like it
already is in tor-rtcompat) so that arti-client users get it too.