ci: Keep doc artifacts
Because Rust 1.91.0 makes it economically unfeasiable to run `cargo doc`
oneselves, it might wise to keep it as an artifact in `rust-dev-recent`
and `rust-nightly`, which generate the docs anyways, so we can keep us
away from having to run the expensive command when reviewing (larger)
changes to documentation.

For now, these artifacts will expire within a week.
Maybe this should get discussed, because on my local machine, they are
over 2GB large?