arti: Change how proxy futures are returned.
Previously, the proxy and dns modules only had "be a proxy"
functions that ran forever.  Now they have functions that bind to
listeners and return a separate "be a proxy" future that runs
forever.

This lets us simplify some kludges in subcommands::proxy.  More
importantly, it will let us return the bound-to ports so that
subcommands::proxy can write them to disk.

This is a break in experimental-apis, which does not require a
semver change.