tor-netdoc: RSA: check exponent and min size in parse2
According to the spec we always use this fixed exponent, and we have a
minimum size of 1024.  This is checked adhoc in the old parser with
some slight assistance from what is now `RsaPublicParse1Helper`.

It's not clear to me that checking the exponent is actually a good
idea.  I think checking the size is probably a bad idea, and if it is
a good idea then 1024 is clearly too short.

But rather than revisit these questions, let's reproduce the old
behaviour in parse2.

In parse2 these checks should be features of the type.