chanmgr: Rename the connect_no_timeout() to be client specific
Heavy refactoring coming up for the ChanBuilder. This renames
connect_no_timeout() to be client specific and uses the client channel
builder directly instead of the generic one.

Previous commit added the outbound ChannelType to the ChanBuilder which
means that we'll soon have a relay specific function to connect using
the relay channel builder.

Refactoring will then happen for shared code in those two functions.

Signed-off-by: David Goulet <dgoulet@torproject.org>