proto: Add traits for public channel views
Add traits that will be returned outside tor-proto allowing us to not
expose client and relay specific channels.

The goal is for the tor-chanmgr to get those objects implementing those
traits and can build and run the reactor without knowing the specific
underlying type.

This allows us to have less code duplication and less client/relay
distinction in the chanmgr.

Signed-off-by: David Goulet <dgoulet@torproject.org>