tor-netdoc: authcert: abolish tmp module
tmp's version of AuthCert is gone.  Each of its fields is already in
AuthCert, including the docs links.  I've decided not to transfer the
syntax snippets.

The remaining function in that module is now an inherent method on
AuthCert, not on tmp's version.

This needs reformatting since verify_self_signed is now at the wrong level!